refactor(web): extract ScanCard + ScrapeCard from admin/scan-page (#688)
`web/src/pages/admin/scan-page.tsx` (532 LOC) bundled two large
self-contained components inline:
- ScanCard (155 LOC)
- ScrapeCard (295 LOC — larger than many whole pages)

Plus ProgressBar (shared) and GroupingStatus, all as local
functions in the same page file.

Moved to their own files under `/features/admin/components/`:
- progress-bar.tsx — 15 LOC shared primitive
- scan-card.tsx — 168 LOC
- scrape-card.tsx — 314 LOC
- grouping-status.tsx — 33 LOC

Page file now reads as its real responsibility: layout + section
wiring. 532 → 24 LOC.

Behaviour preserved; all 1557 web tests pass. `npx tsc --noEmit`
clean.

Deferred (mentioned in #688): the shared "active / complete / error"
OperationStatusCard abstraction. ScanCard and ScrapeCard have
enough divergent content in each phase (different stats in active,
different counts in complete) that the shared card would need 3+
flexible slots. Worth a follow-up if a 3rd operation card arrives.

Closes #688
